move test to new home of corresponding func
